refactor(timeline): extract context-menu state into useContextMenuState
Move closerEdge state + handleContextMenu out of TimelineItem into a
small focused hook. The larger pointer handlers (handleClick,
handleMouseDown, handleDoubleClick, handleSmartTrimStart) stay in the
host file because they are tightly coupled to local trim/stretch/slip
state and consume too many host-level dependencies to extract without
introducing churn in their dep arrays.

Pull-out is intentionally minimal here: the bigger sub-component split
in the next change does the rest of the de-coupling work.
